fix libffi for 32-bit centos

+++ config/software/libffi.rb
@@ -62,9 +62,9 @@ build do
   # libffi's default install location of header files is aweful...
   command "cp -f #{install_dir}/embedded/lib/libffi-3.0.13/include/* #{install_dir}/embedded/include"
 
-  # On centos libffi libraries are places under /embedded/lib64
+  # On 64-bit centos, libffi libraries are places under /embedded/lib64
   # move them over to lib
-  if platform == "centos"
+  if platform == "centos" && _64_bit?
     command "mv #{install_dir}/embedded/lib64/* #{install_dir}/embedded/lib/"
     command "rm -rf #{install_dir}/embedded/lib64"
   end
